Fix memory leak in unbounded storage case
Evictions were no being tracked properly when maxSize triggered an
eviction.

In the max-bound storage case, this wasn't so much of a problem, because
the storage couldn't grow unbounded anyway.  However, since size-evicted
indexes were not being added to the `free` list, it would effectively
reduce the available storage space, holding onto old entries and
shortening the available storage area, which would result in fewer items
being cached (except some that were cached indefinitely!)

In the unbounded case, this was much worse.  Because the storage array
was unbounded, the cache effectively always remained in the initial
"fill to completion" state, resulting in holding onto _every_ item
forever, while thinking that the `calculatedSize` was being limited to
`maxSize`!  Pretty bad.

(This will be back-ported to other 7.x versions as a patch.)

Thanks to @moish83 for the report.

Fix: https://github.com/isaacs/node-lru-cache/issues/227