handle failures from fetchMethod
Remove the failed fetch key from the cache, rather than raising an
unhandled rejection error when the promise was not returned.

At present, there is no way to suppress this behavior, but it is
still preferable to raising an unhandled exception in these cases.

Fix: https://github.com/isaacs/node-lru-cache/issues/233