perf, type check fetch context, rename to context
This streamlines a few small performance drags, but the main change is
that 'fetchContext' is renamed to just 'context', is strongly typed
(though it defaults to 'unknown'), and is no longer defaulted on the
constructor.

When the type is defined in the constructor generic to something other
than unknown, void, or undefined, it is required on the calls to
fetch().  If defined to void or undefined, it is not allowed in the
calls to fetch.