fix infinite loop when using .get() inside .keys() loop
This fixes an issue where you can enter an infinite loop if you use
`cache.get()` whilst iterating over `cache.keys()`. As far as
I understand the infinite loop occurs because the index order changes
when you call `cache.get()`. This can be avoided by using `cache.peek()`
but is still unexpected. To resolve this, inside the `.keys` method I've
added a `Set` that stores visited keys, and modified the logic to not
yield a key if it has already been visited.
This avoids the case where the index value changes during iteration, but
allows iteration to include keys added whilst iterating.

Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>