Add an internal function to refresh timer used for ttl
When overriding the global timers with a custom timer, the timer
used for ttl is not updated and will continue to use the timer that
was defined when the package was first imported.

This adds an internal function that can be called to refresh the
timer used for ttl with the current global timer.