add 'perf' option
This enables the consumer to override Date/performance object usage once
at the point of LRU object creation.

This is an affordance to make the TTL behavior easier to test in some
scenarios, without sacrificing security or performance. It should most
likely not be used in production!

Re: #345

PR-URL: https://github.com/isaacs/node-lru-cache/pull/386
Credit: @isaacs
Close: #386
Reviewed-by: @isaacs