clean up Timer objects when using ttlAutopurge
This prevents a possible runaway resource utilization issue when using
ttlAutopurge.

If a single key is written to repetedly, then many NodeJS.Timer objects
will be created (or equivalent in the browser), and they are not
unscheduled when no longer needed. This results in many checks for
staleness happening unnecessarily, and so on.

Now, when a key is overwritten, set with no TTL, or any other behavior
that will mean that autopurge is no longer necessary, the timer is
cleared and deleted.